refactor: these two pages too
